Parse FIXNUM as Long (not as Integer).
If a FIXNUM is too large for an Integer a NumberFormatException will
be thrown. Therefore use Long instead.

+++ do_jdbc/src/main/java/data_objects/drivers/AbstractDriverDefinition.java
@@ -439,7 +439,7 @@ public abstract class AbstractDriverDefinition implements DriverDefinition {
             IRubyObject arg, int idx) throws SQLException {
         switch (RubyType.inferRubyType(arg)) {
         case FIXNUM:
-            ps.setInt(idx, Integer.parseInt(arg.toString()));
+            ps.setLong(idx, Long.parseLong(arg.toString()));
             break;
         case BIGNUM:
             BigInteger big = ((RubyBignum) arg).getValue();
